MTCP: Don't save zero pages to ckpt-image.
Detect zero pages and do not write them to ckpt image. Instead, set a
special flag, MTCP_PROT_ZERO_PAGE, in Area.prot. On restart, we detect
this MTCP_PROT_ZERO_PAGE flag and map the area as MAP_ANONYMOUS. This
way, these pages will be backed by shared a zero-page and will reduce
the resident memory footprint of the process.

+++ mtcp/mtcp_util.c
@@ -286,8 +286,8 @@ void mtcp_readfile(int fd, void *buf, size_t size)
       mtcp_abort();
     }
     else if (rc == 0) {
-      MTCP_PRINTF("only read %zu bytes instead of %zu from checkpoint file\n",
-                  ar, size);
+      MTCP_PRINTF("only read %u bytes instead of %u from checkpoint file\n",
+                  (unsigned)ar, (unsigned)size);
       if (tries++ >= 10) {
         MTCP_PRINTF(" failed to read after 10 tries in a row.\n");
         mtcp_abort();
